# contracts/source/openapi.yaml
openapi: 3.1.0
info:
  title: SourceHarbor API
  version: 0.1.19
  description: API contract for subscriptions, ingest polling, jobs, videos and artifacts.
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/v1/reader/batches/{batch_id}/judge:
    post:
      operationId: judgeConsumptionBatch
      parameters:
        - in: path
          name: batch_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Cluster verdict manifest generated for the consumption batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterVerdictManifest'
        '400':
          description: Invalid batch state
        '404':
          description: Not found
  /api/v1/reader/batches/{batch_id}/manifest:
    get:
      operationId: getClusterVerdictManifest
      parameters:
        - in: path
          name: batch_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cluster verdict manifest detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterVerdictManifest'
        '404':
          description: Not found
  /api/v1/reader/batches/{batch_id}/materialize:
    post:
      operationId: materializeConsumptionBatch
      parameters:
        - in: path
          name: batch_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterializeBatchRequest'
      responses:
        '201':
          description: Published reader documents generated for the consumption batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderBatchMaterializationResponse'
        '400':
          description: Invalid batch state
        '404':
          description: Not found
  /api/v1/reader/documents:
    get:
      operationId: listPublishedReaderDocuments
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: window_id
          schema:
            type: string
      responses:
        '200':
          description: Current published reader documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedReaderDocumentListResponse'
  /api/v1/reader/documents/slug/{slug}:
    get:
      operationId: getPublishedReaderDocumentBySlug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Published reader document detail by slug
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedReaderDocument'
        '404':
          description: Not found
  /api/v1/reader/documents/{document_id}:
    get:
      operationId: getPublishedReaderDocument
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Published reader document detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedReaderDocument'
        '404':
          description: Not found
  /api/v1/reader/documents/{document_id}/repair:
    post:
      operationId: repairPublishedReaderDocument
      parameters:
        - in: path
          name: document_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReaderDocumentRepairRequest'
      responses:
        '201':
          description: Published reader document repaired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedReaderDocument'
        '400':
          description: Invalid repair request
        '404':
          description: Not found
  /api/v1/reader/navigation-brief:
    get:
      operationId: getReaderNavigationBrief
      parameters:
        - in: query
          name: window_id
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 8
      responses:
        '200':
          description: Navigation brief over current published reader documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigationBrief'
...
components:
  schemas:
...
    ClusterVerdictManifestMember:
      type: object
      required:
        [source_item_id, platform, source_origin, title, claim_kinds, digest_preview]
      properties:
        source_item_id:
          type: string
        job_id:
          type: string
          nullable: true
        platform:
          type: string
        source_origin:
          type: string
        title:
          type: string
        source_url:
          type: string
          nullable: true
        published_at:
          type: string
          nullable: true
        claim_kinds:
          type: array
          items:
            type: string
        digest_preview:
          type: string
    ClusterVerdictCluster:
      type: object
      required:
        [
          cluster_id,
          cluster_key,
          topic_label,
          decision,
          source_item_count,
          source_item_ids,
          job_ids,
          platforms,
          claim_kinds,
          headline,
          digest_preview,
          members
        ]
      properties:
        cluster_id:
          type: string
        cluster_key:
          type: string
        topic_key:
          type: string
          nullable: true
        topic_label:
          type: string
        decision:
          type: string
          enum: [merge_then_polish]
        source_item_count:
          type: integer
          minimum: 0
        source_item_ids:
          type: array
          items:
            type: string
        job_ids:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
        claim_kinds:
          type: array
          items:
            type: string
        headline:
          type: string
        digest_preview:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/ClusterVerdictManifestMember'
    ClusterVerdictSingleton:
      type: object
      required:
        [
          singleton_id,
          source_item_id,
          platform,
          source_origin,
          content_type,
          title,
          claim_kinds,
          decision,
          digest_preview
        ]
      properties:
        singleton_id:
          type: string
        source_item_id:
          type: string
        ingest_run_item_id:
          type: string
          nullable: true
        job_id:
          type: string
          nullable: true
        platform:
          type: string
        source_origin:
          type: string
        content_type:
          type: string
        title:
          type: string
        source_url:
          type: string
          nullable: true
        published_at:
          type: string
          nullable: true
        topic_key:
          type: string
          nullable: true
        topic_label:
          type: string
          nullable: true
        claim_kinds:
          type: array
          items:
            type: string
        decision:
          type: string
          enum: [polish_only]
        digest_preview:
          type: string
    ClusterVerdictManifestPayload:
      type: object
      required:
        [
          manifest_kind,
          generated_at,
          consumption_batch_id,
          window_id,
          status,
          source_item_count,
          cluster_count,
          singleton_count,
          clusters,
          singletons
        ]
      properties:
        manifest_kind:
          type: string
        generated_at:
          type: string
        consumption_batch_id:
          type: string
        window_id:
          type: string
        status:
          type: string
          enum: [ready, gap_detected]
        source_item_count:
          type: integer
          minimum: 0
        cluster_count:
          type: integer
          minimum: 0
        singleton_count:
          type: integer
          minimum: 0
        clusters:
          type: array
          items:
            $ref: '#/components/schemas/ClusterVerdictCluster'
        singletons:
          type: array
          items:
            $ref: '#/components/schemas/ClusterVerdictSingleton'
    ClusterVerdictManifest:
      type: object
      required:
        [
          id,
          consumption_batch_id,
          window_id,
          status,
          source_item_count,
          cluster_count,
          singleton_count,
          manifest,
          created_at,
          updated_at
        ]
      properties:
        id:
          type: string
          format: uuid
        consumption_batch_id:
          type: string
          format: uuid
        window_id:
          type: string
        status:
          type: string
          enum: [ready, gap_detected]
        source_item_count:
          type: integer
          minimum: 0
        cluster_count:
          type: integer
          minimum: 0
        singleton_count:
          type: integer
          minimum: 0
        summary_markdown:
          type: string
          nullable: true
        manifest:
          $ref: '#/components/schemas/ClusterVerdictManifestPayload'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReaderDocumentSection:
      type: object
      required: [section_id, title, markdown, source_item_ids]
      properties:
        section_id:
          type: string
        title:
          type: string
        markdown:
          type: string
        source_item_ids:
          type: array
          items:
            type: string
    ReaderDocumentWarning:
      type: object
      required: [warning_kind, published_with_gap, reasons]
      properties:
        warning_kind:
          type: string
        warning_kinds:
          type: array
          items:
            type: string
        published_with_gap:
          type: boolean
        reasons:
          type: array
          items:
            type: string
        failed_source_count:
          type: integer
          minimum: 0
        degraded_source_count:
          type: integer
          minimum: 0
        missing_digest_count:
          type: integer
          minimum: 0
        affected_scope:
          type: object
          additionalProperties: true
        version:
          type: integer
          nullable: true
        readable_why:
          type: string
          nullable: true
        not_fully_sealed_why:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        generated_at:
          type: string
          nullable: true
    PublishedReaderDocument:
      type: object
      required:
        [
          id,
          stable_key,
          slug,
          window_id,
          title,
          markdown,
          materialization_mode,
          version,
          publish_status,
          published_with_gap,
          is_current,
          source_item_count,
          warning,
          coverage_ledger,
          traceability_pack,
          source_refs,
          sections,
          repair_history,
          created_at,
          updated_at
        ]
      properties:
        id:
          type: string
          format: uuid
        stable_key:
          type: string
        slug:
          type: string
        window_id:
          type: string
        topic_key:
          type: string
          nullable: true
        topic_label:
          type: string
          nullable: true
        title:
          type: string
        summary:
          type: string
          nullable: true
        markdown:
          type: string
        materialization_mode:
          type: string
          enum:
            [merge_then_polish, polish_only, repair_patch, repair_section, repair_cluster]
        version:
          type: integer
          minimum: 1
        publish_status:
          type: string
          enum: [published, published_with_gap, superseded]
        published_with_gap:
          type: boolean
        is_current:
          type: boolean
        source_item_count:
          type: integer
          minimum: 0
        consumption_batch_id:
          type: string
          nullable: true
        cluster_verdict_manifest_id:
          type: string
          nullable: true
        supersedes_document_id:
          type: string
          nullable: true
        warning:
          $ref: '#/components/schemas/ReaderDocumentWarning'
        coverage_ledger:
          type: object
        traceability_pack:
          type: object
        source_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceIdentityRef'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ReaderDocumentSection'
        repair_history:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReaderBatchMaterializationResponse:
      type: object
      required:
        [
          consumption_batch_id,
          cluster_verdict_manifest_id,
          window_id,
          published_document_count,
          published_with_gap_count,
          documents,
          navigation_brief
        ]
      properties:
        consumption_batch_id:
          type: string
          format: uuid
        cluster_verdict_manifest_id:
          type: string
        window_id:
          type: string
        published_document_count:
          type: integer
          minimum: 0
        published_with_gap_count:
          type: integer
          minimum: 0
        documents:
          type: array
          items:
            $ref: '#/components/schemas/PublishedReaderDocument'
        navigation_brief:
          $ref: '#/components/schemas/NavigationBrief'
    ReaderDocumentRepairRequest:
      type: object
      required: [repair_mode]
      properties:
        repair_mode:
          type: string
          enum: [patch, section, cluster]
        section_ids:
          type: array
          items:
            type: string
    MaterializeBatchRequest:
      type: object
      properties:
        reader_output_locale:
          type: string
          default: zh-CN
        reader_style_profile:
          type: string
          default: briefing
    PublishedReaderDocumentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublishedReaderDocument'
    NavigationBrief:
      type: object
      required:
        [
          brief_kind,
          generated_at,
          window_id,
          document_count,
          published_with_gap_count,
          summary,
          items
        ]
      properties:
        brief_kind:
          type: string
        generated_at:
          type: string
        window_id:
          type: string
        document_count:
          type: integer
          minimum: 0
        published_with_gap_count:
          type: integer
          minimum: 0
        summary:
          type: string
        items:
          type: array
          items:
            type: object
...
    SourceIdentityRef:
      type: object
      required:
        [source_item_id, platform, source_origin, title, claim_kinds, digest_preview]
      properties:
        source_item_id:
          type: string
        job_id:
          type: string
          nullable: true
        platform:
          type: string
        source_origin:
          type: string
        title:
          type: string
        source_url:
          type: string
          nullable: true
        published_at:
          type: string
          nullable: true
        claim_kinds:
          type: array
          items:
            type: string
        digest_preview:
          type: string
        subscription_id:
          type: string
          nullable: true
        matched_subscription_name:
          type: string
          nullable: true
        relation_kind:
          type: string
          nullable: true
        affiliation_label:
          type: string
          nullable: true
        canonical_source_name:
          type: string
          nullable: true
        canonical_author_name:
          type: string
          nullable: true
        creator_display_name:
          type: string
          nullable: true
        creator_handle:
          type: string
          nullable: true
        thumbnail_url:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        avatar_label:
          type: string
          nullable: true
        identity_status:
          type: string
          nullable: true
        raw_stage_contract:
          $ref: '#/components/schemas/RawStageContractReceipt'
        job_bundle_route:
          type: string
          nullable: true
    RawStageContractReceipt:
      type: object
      properties:
        content_type:
          type: string
          nullable: true
        analysis_mode:
          type: string
          nullable: true
        video_first:
          type: boolean
          nullable: true
        video_input_required:
          type: boolean
          nullable: true
        preprocess_enabled:
          type: boolean
          nullable: true
        preprocess_model:
          type: string
          nullable: true
        preprocess_input_mode:
          type: string
          nullable: true
        preprocess_media_input:
          type: string
          nullable: true
        review_required:
          type: boolean
          nullable: true
        review_model:
          type: string
          nullable: true
        review_input_mode:
          type: string
          nullable: true
        primary_model:
          type: string
          nullable: true
        primary_input_mode:
          type: string
          nullable: true
        primary_media_input:
          type: string
          nullable: true
        review_media_input:
          type: string
          nullable: true
        video_contract_satisfied:
          type: boolean
          nullable: true
...